Unshim getGpuColumnarToRowTransition
